+++ elastic-job-cloud/elastic-job-cloud-scheduler/src/test/java/com/dangdang/ddframe/job/cloud/scheduler/context/AllContextTests.java
@@ -21,8 +21,6 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({
-        JobContextTest.class
-    })
+@Suite.SuiteClasses(JobContextTest.class)
 public final class AllContextTests {
 }