+++ elastic-job-common/elastic-job-common-core/src/main/java/com/dangdang/ddframe/job/executor/handler/impl/DefaultExecutorServiceHandler.java
@@ -30,7 +30,7 @@ import java.util.concurrent.ExecutorService;
 public final class DefaultExecutorServiceHandler implements ExecutorServiceHandler {
     
     @Override
-    public ExecutorService createExecutorService() {
-        return new ExecutorServiceObject("inner-job", Runtime.getRuntime().availableProcessors() * 2).createExecutorService();
+    public ExecutorService createExecutorService(final String jobName) {
+        return new ExecutorServiceObject("inner-job-" + jobName, Runtime.getRuntime().availableProcessors() * 2).createExecutorService();
     }
 }