+++ elastic-job-lite/elastic-job-lite-spring/src/main/java/com/dangdang/ddframe/job/lite/spring/schedule/SpringJobScheduler.java
@@ -21,6 +21,7 @@ import com.dangdang.ddframe.job.api.ElasticJob;
 import com.dangdang.ddframe.job.event.JobEventConfiguration;
 import com.dangdang.ddframe.job.lite.api.JobScheduler;
 import com.dangdang.ddframe.job.lite.api.listener.ElasticJobListener;
+import com.dangdang.ddframe.job.lite.config.LiteJobConfiguration;
 import com.dangdang.ddframe.job.lite.spring.namespace.parser.common.AbstractJobConfigurationDto;
 import com.dangdang.ddframe.job.lite.spring.util.AopTargetUtils;
 import com.dangdang.ddframe.job.reg.base.CoordinatorRegistryCenter;
@@ -34,9 +35,9 @@ import com.google.common.base.Optional;
  */
 public class SpringJobScheduler extends JobScheduler {
     
-    private final ElasticJob elasticJob; 
+    private final ElasticJob elasticJob;
     
-    public SpringJobScheduler(final ElasticJob elasticJob, final CoordinatorRegistryCenter regCenter, final AbstractJobConfigurationDto jobConfigDto, final ElasticJobListener[]elasticJobListeners) {
+    public SpringJobScheduler(final ElasticJob elasticJob, final CoordinatorRegistryCenter regCenter, final AbstractJobConfigurationDto jobConfigDto, final ElasticJobListener[] elasticJobListeners) {
         super(regCenter, jobConfigDto.toLiteJobConfiguration(), getTargetElasticJobListeners(elasticJobListeners));
         this.elasticJob = elasticJob;
     }
@@ -44,6 +45,17 @@ public class SpringJobScheduler extends JobScheduler {
     public SpringJobScheduler(final ElasticJob elasticJob, final CoordinatorRegistryCenter regCenter, final AbstractJobConfigurationDto jobConfigDto, 
                               final JobEventConfiguration jobEventConfig, final ElasticJobListener[]elasticJobListeners) {
         super(regCenter, jobConfigDto.toLiteJobConfiguration(), jobEventConfig, getTargetElasticJobListeners(elasticJobListeners));
+        this.elasticJob = elasticJob;
+    }
+    
+    public SpringJobScheduler(final ElasticJob elasticJob, final CoordinatorRegistryCenter regCenter, final LiteJobConfiguration jobConfig, final ElasticJobListener[] elasticJobListeners) {
+        super(regCenter, jobConfig, getTargetElasticJobListeners(elasticJobListeners));
+        this.elasticJob = elasticJob;
+    }
+    
+    public SpringJobScheduler(final ElasticJob elasticJob, final CoordinatorRegistryCenter regCenter, final LiteJobConfiguration jobConfig, 
+                              final JobEventConfiguration jobEventConfig, final ElasticJobListener[] elasticJobListeners) {
+        super(regCenter, jobConfig, jobEventConfig, getTargetElasticJobListeners(elasticJobListeners));
         this.elasticJob = elasticJob;
     }
     
@@ -57,6 +69,6 @@ public class SpringJobScheduler extends JobScheduler {
     
     @Override
     protected Optional<ElasticJob> createElasticJobInstance() {
-        return Optional.of(elasticJob);
+        return Optional.fromNullable(elasticJob);
     }
 }